COMMISS.DYNAP.MNT: Move Dynap Server to same module as RCDS optimizer.
